% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diggs-package.R
\docType{package}
\name{diggs-package}
\alias{diggs}
\alias{diggs-package}
\title{diggs: Delineation-Informed Geographic Ground Selection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Interactive Shiny app for selecting historic orthophoto imagery from the BC Data Catalogue. Draw polygons or upload a custom AOI, filter by year/media/scale, view estimated photo footprints, and export selections to CSV. Built on the fly package for spatial photo operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NewGraphEnvironment/diggs}
  \item Report bugs at \url{https://github.com/NewGraphEnvironment/diggs/issues}
}

}
\author{
\strong{Maintainer}: Al Irvine \email{al@newgraphenvironment.com}

Other contributors:
\itemize{
  \item New Graph Environment Ltd. [copyright holder]
}

}
\keyword{internal}
